requirements.txt
alembic==1.4.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
Mako==1.1.3
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.4
pandas==1.1.5
psycopg2-binary==2.8.6
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
searchtweets-v2==1.0.4
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
tqdm==4.54.1
urllib3==1.26.2
Werkzeug==1.0.1
